requirements.txt
python-docx==1.1.2
pandas==2.2.3
tk==0.1.0
ttkbootstrap==1.10.1
